Create locked file on Windows
