+++ go/libkb/socket.go
@@ -1,5 +1,3 @@
-// +build darwin dragonfly freebsd linux nacl netbsd openbsd solaris
-
 package libkb
 
 import (
